add more icons fix message boxes default button, etc. fix select cover and core confirmation dialogs add category to hmods readme

+++ SelectSystemDialog.cs
@@ -51,28 +51,31 @@ namespace com.clusterrr.hakchi_gui
 
         private void buttonPrepare_Click(object sender, EventArgs e)
         {
-            for(int i = 0; i < checkedListBox.Items.Count; ++i)
+            if (Tasks.MessageForm.Show(this, Resources.ApplyChanges, Resources.ApplyChangesQ, Resources.sign_question, new Tasks.MessageForm.Button[] { Tasks.MessageForm.Button.Yes, Tasks.MessageForm.Button.No }, Tasks.MessageForm.DefaultButton.Button1) == Tasks.MessageForm.Button.Yes)
             {
-                var path = Shared.PathCombine(Program.BaseDirectoryExternal, "art", checkedListBox.Items[i] as string);
-                if (checkedListBox.GetItemChecked(i))
+                for (int i = 0; i < checkedListBox.Items.Count; ++i)
                 {
-                    Directory.CreateDirectory(path);
-                }
-                else
-                {
-                    if (Directory.Exists(path))
+                    var path = Shared.PathCombine(Program.BaseDirectoryExternal, "art", checkedListBox.Items[i] as string);
+                    if (checkedListBox.GetItemChecked(i))
                     {
-                        try
+                        Directory.CreateDirectory(path);
+                    }
+                    else
+                    {
+                        if (Directory.Exists(path))
                         {
-                            // this will only delete empty directories, which is the desired outcome
-                            Directory.Delete(path);
+                            try
+                            {
+                                // this will only delete empty directories, which is the desired outcome
+                                Directory.Delete(path);
+                            }
+                            catch { }
                         }
-                        catch { }
                     }
                 }
+                if (!ConfigIni.Instance.DisablePopups)
+                    Tasks.MessageForm.Show(Resources.Wow, Resources.Done, Resources.sign_check);
             }
-            if (!ConfigIni.Instance.DisablePopups)
-                Tasks.MessageForm.Show(Resources.Wow, Resources.Done, Resources.sign_check);
             Close();
         }
     }
